add obfuscated string hack to pb, use it in middleware configs
