+++ config/webhook_verification.go
@@ -1,6 +1,8 @@
 package config
 
-import "golang.ngrok.com/ngrok/internal/pb"
+import (
+	"golang.ngrok.com/ngrok/internal/pb"
+)
 
 // Configuration for webhook verification.
 type webhookVerification struct {
@@ -16,7 +18,7 @@ func (wv *webhookVerification) toProtoConfig() *pb.MiddlewareConfiguration_Webho
 	}
 	return &pb.MiddlewareConfiguration_WebhookVerification{
 		Provider: wv.Provider,
-		Secret:   wv.Secret,
+		Secret:   pb.ObfuscatedString(wv.Secret),
 	}
 }
