+++ internal/tunnel/proto/msg_test.go
@@ -14,6 +14,9 @@ func TestObfuscatedString(t *testing.T) {
 	obfuscatedString := ObfuscatedString(fakeToken)
 	require.Equal(t, fakeToken, obfuscatedString.PlainText())
 
+	// Intentionally doing something a little funky here to make
+	// sure that the String() method is doing what it's supposed to.
+	//nolint:gosimple
 	printedString := fmt.Sprintf("%s", obfuscatedString)
 	require.NotEqual(t, fakeToken, printedString)
 	require.Equal(t, "HIDDEN", printedString)
